requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-google-genai
langchain-community
langchain-text-splitters
langchain-chroma
langchain-openai
pypdf